app/requirements.txt
black==25.12.0
Django==6.0
django-extensions==4.1
djangorestframework==3.16.1
flake8==7.3.0
isort==7.0.0
psycopg2-binary==2.9.11
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
factory-boy==3.3.1